synthetic_data/requirements.txt
aiofiles
requests
pillow
opencv-python
openai
dirtyjson
